Complete packet system

+++ src/main/java/com/xumuk/realism/packets/YearSyncMessage.java
@@ -9,21 +9,21 @@ import net.minecraft.entity.player.EntityPlayerMP;
 public class YearSyncMessage extends SRSimplePacket {
 	public YearSyncMessage() {}
 
-	public YearSyncMessage(byte year) {
-		buf().writeByte(year);
+	public YearSyncMessage(int year) {
+		buf().writeInt(year);
 	}
 
 	@Override
 	public void client(EntityPlayer player) {
 		if (player != null) {
 			IDate cap = player.getEntityWorld().getCapability(DateProvider.DATE, null);
-			if (cap != null) { cap.setYear(buf().readByte()); }
+			if (cap != null) { cap.setYear(buf().readInt()); }
 		}
 	}
 
 	@Override
 	public void server(EntityPlayerMP player) {
 		IDate date = player.getEntityWorld().getCapability(DateProvider.DATE, null);
-		if (date != null) date.setYear(buf().readByte());
+		if (date != null) date.setYear(buf().readInt());
 	}
 }
